response packet append class somewhat working

+++ includes/HttpSocket.hpp
@@ -38,7 +38,7 @@ private:
     // private functions
     sockaddr_in _compose_sockaddr(const std::string &addr, int port);
     std::unique_ptr<TcpSocket> _create_bind_socket(const sockaddr_in &address);
-    int _smallest_max_body_size() const;
+    size_t _smallest_max_body_size() const;
     void _write_client_response(std::unique_ptr<ResponsePacket> response);
 };
